IGNITE-19646 after code review #1.0
